- Menu up left works

+++ app/src/main/java/com/wua/mc/webuntisapp/model/DataBaseObject.java
@@ -10,8 +10,8 @@ public class DataBaseObject {
 
     private long event_id;
     private String event_room;
-    private int event_timestamp_start;
-    private int event_timestamp_end;
+    private long event_timestamp_start;
+    private long event_timestamp_end;
     private String event_name;
     private String event_color;
     private String event_type;
@@ -29,7 +29,7 @@ public class DataBaseObject {
 
     }
 
-    public DataBaseObject(String event_room, int event_timestamp_start, int event_timestamp_end, String event_name, String event_color, String event_type, long event_id){
+    public DataBaseObject(String event_room, long event_timestamp_start, long event_timestamp_end, String event_name, String event_color, String event_type, long event_id){
 
         this.setEvent_id(event_id);
         this.setEvent_room(event_room);
@@ -74,11 +74,11 @@ public class DataBaseObject {
         this.event_room = event_room;
     }
 
-    public void setEvent_timestamp_start(int event_timestamp_start) {
+    public void setEvent_timestamp_start(long event_timestamp_start) {
         this.event_timestamp_start = event_timestamp_start;
     }
 
-    public void setEvent_timestamp_end(int event_timestamp_end) {
+    public void setEvent_timestamp_end(long event_timestamp_end) {
         this.event_timestamp_end = event_timestamp_end;
     }
 
@@ -127,11 +127,11 @@ public class DataBaseObject {
         return event_room;
     }
 
-    public int getEvent_timestamp_start() {
+    public long getEvent_timestamp_start() {
         return event_timestamp_start;
     }
 
-    public int getEvent_timestamp_end() {
+    public long getEvent_timestamp_end() {
         return event_timestamp_end;
     }
